requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
pandas==2.2.2
numpy==1.26.4
duckdb==1.1.0
matplotlib==3.9.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
python-multipart==0.0.9
pyyaml==6.0.2
scipy==1.13.1
